fix: move structs to separate file
